Fix cast bug in scale_factor

+++ source/src/node/fft/ifft.cpp
@@ -126,7 +126,7 @@ void IFFT::process(sample **out, int num_frames)
      *-----------------------------------------------------------------------*/
     for (int hop = 0; hop < this->num_hops; hop++)
     {
-        float scale_factor = hop_size / fft_size;
+        float scale_factor = (float) hop_size / fft_size;
         this->ifft(this->input->out[hop],
                    this->out[0] + (hop * hop_size),
                    true,
